fixing bug on import layout on menu bar
