+++ packages/suite-base/src/components/LayoutBrowser/index.tsx
@@ -20,9 +20,7 @@ import {
 import * as _ from "lodash-es";
 import moment from "moment";
 import { useSnackbar } from "notistack";
-import path from "path";
-import { MouseEvent, useCallback, useEffect, useLayoutEffect, useMemo } from "react";
-import { useMountedState } from "react-use";
+import { useEffect, useLayoutEffect, useMemo } from "react";
 import useAsyncFn from "react-use/lib/useAsyncFn";
 import { makeStyles } from "tss-react/mui";
 
@@ -36,7 +34,6 @@ import { useAnalytics } from "@lichtblick/suite-base/context/AnalyticsContext";
 import {
   LayoutID,
   LayoutState,
-  useCurrentLayoutActions,
   useCurrentLayoutSelector,
 } from "@lichtblick/suite-base/context/CurrentLayoutContext";
 import { LayoutData } from "@lichtblick/suite-base/context/CurrentLayoutContext/actions";
@@ -45,12 +42,13 @@ import { useLayoutManager } from "@lichtblick/suite-base/context/LayoutManagerCo
 import { useAppConfigurationValue } from "@lichtblick/suite-base/hooks/useAppConfigurationValue";
 import useCallbackWithToast from "@lichtblick/suite-base/hooks/useCallbackWithToast";
 import { useConfirm } from "@lichtblick/suite-base/hooks/useConfirm";
+import { useLayoutActions } from "@lichtblick/suite-base/hooks/useLayoutActions";
+import { useLayoutNavigation } from "@lichtblick/suite-base/hooks/useLayoutNavigation";
+import { useLayoutTransfer } from "@lichtblick/suite-base/hooks/useLayoutTransfer";
 import { usePrompt } from "@lichtblick/suite-base/hooks/usePrompt";
 import { defaultPlaybackConfig } from "@lichtblick/suite-base/providers/CurrentLayoutProvider/reducers";
 import { AppEvent } from "@lichtblick/suite-base/services/IAnalytics";
 import { Layout, layoutIsShared } from "@lichtblick/suite-base/services/ILayoutStorage";
-import { downloadTextFile } from "@lichtblick/suite-base/util/download";
-import showOpenFilePicker from "@lichtblick/suite-base/util/showOpenFilePicker";
 
 import LayoutSection from "./LayoutSection";
 import { useLayoutBrowserReducer } from "./reducer";
@@ -66,7 +64,6 @@ const useStyles = makeStyles()((theme) => ({
 }));
 
 export default function LayoutBrowser({
-  menuClose,
   currentDateForStorybook,
 }: React.PropsWithChildren<{
   menuClose?: () => void;
@@ -74,16 +71,15 @@ export default function LayoutBrowser({
 }>): React.JSX.Element {
   const { classes } = useStyles();
   const { signIn } = useCurrentUser();
-  const isMounted = useMountedState();
   const { enqueueSnackbar } = useSnackbar();
   const layoutManager = useLayoutManager();
   const [prompt, promptModal] = usePrompt();
   const analytics = useAnalytics();
-  const [confirm, confirmModal] = useConfirm();
-  const { unsavedChangesPrompt, openUnsavedChangesPrompt } = useUnsavedChangesPrompt();
+  //PROBLEM HERE WITH IMPORT
+  const [, confirmModal] = useConfirm();
+  const { unsavedChangesPrompt } = useUnsavedChangesPrompt();
 
   const currentLayoutId = useCurrentLayoutSelector(selectedLayoutIdSelector);
-  const { setSelectedLayoutId } = useCurrentLayoutActions();
 
   const [state, dispatch] = useLayoutBrowserReducer({
     lastSelectedId: currentLayoutId,
@@ -91,6 +87,12 @@ export default function LayoutBrowser({
     error: layoutManager.error,
     online: layoutManager.isOnline,
   });
+
+  const { onRenameLayout, onDuplicateLayout, onDeleteLayout, onRevertLayout, onOverwriteLayout } =
+    useLayoutActions();
+  const { importLayout, exportLayout } = useLayoutTransfer();
+  const { promptForUnsavedChanges, onSelectLayout } = useLayoutNavigation();
+  const onExportLayout = exportLayout;
 
   useLayoutEffect(() => {
     const busyListener = () => {
@@ -194,161 +196,6 @@ export default function LayoutBrowser({
     });
   }, [reloadLayouts]);
 
-  /**
-   * Don't allow the user to switch away from a personal layout if they have unsaved changes. This
-   * currently has a race condition because of the throttled save in CurrentLayoutProvider -- it's
-   * possible to make changes and switch layouts before they're sent to the layout manager.
-   * @returns true if the original action should continue, false otherwise
-   */
-  const promptForUnsavedChanges = useCallback(async () => {
-    const currentLayout =
-      currentLayoutId != undefined ? await layoutManager.getLayout(currentLayoutId) : undefined;
-    if (
-      currentLayout != undefined &&
-      layoutIsShared(currentLayout) &&
-      currentLayout.working != undefined
-    ) {
-      const result = await openUnsavedChangesPrompt(currentLayout);
-      switch (result.type) {
-        case "cancel":
-          return false;
-        case "discard":
-          await layoutManager.revertLayout({ id: currentLayout.id });
-          void analytics.logEvent(AppEvent.LAYOUT_REVERT, {
-            permission: currentLayout.permission,
-            context: "UnsavedChangesPrompt",
-          });
-          return true;
-        case "overwrite":
-          await layoutManager.overwriteLayout({ id: currentLayout.id });
-          void analytics.logEvent(AppEvent.LAYOUT_OVERWRITE, {
-            permission: currentLayout.permission,
-            context: "UnsavedChangesPrompt",
-          });
-          return true;
-        case "makePersonal":
-          // We don't use onMakePersonalCopy() here because it might need to prompt for unsaved changes, and we don't want to select the newly created layout
-          await layoutManager.makePersonalCopy({
-            id: currentLayout.id,
-            name: result.name,
-          });
-          void analytics.logEvent(AppEvent.LAYOUT_MAKE_PERSONAL_COPY, {
-            permission: currentLayout.permission,
-            syncStatus: currentLayout.syncInfo?.status,
-            context: "UnsavedChangesPrompt",
-          });
-          return true;
-      }
-    }
-    return true;
-  }, [analytics, currentLayoutId, layoutManager, openUnsavedChangesPrompt]);
-
-  const onSelectLayout = useCallbackWithToast(
-    async (
-      item: Layout,
-      { selectedViaClick = false, event }: { selectedViaClick?: boolean; event?: MouseEvent } = {},
-    ) => {
-      if (selectedViaClick) {
-        if (!(await promptForUnsavedChanges())) {
-          return;
-        }
-        void analytics.logEvent(AppEvent.LAYOUT_SELECT, { permission: item.permission });
-      }
-      if (event?.ctrlKey === true || event?.metaKey === true || event?.shiftKey === true) {
-        if (item.id !== currentLayoutId) {
-          dispatch({
-            type: "select-id",
-            id: item.id,
-            layouts: layouts.value,
-            modKey: event.ctrlKey || event.metaKey,
-            shiftKey: event.shiftKey,
-          });
-        }
-      } else {
-        setSelectedLayoutId(item.id);
-        dispatch({ type: "select-id", id: item.id });
-        menuClose?.();
-      }
-    },
-    [
-      analytics,
-      currentLayoutId,
-      dispatch,
-      layouts.value,
-      menuClose,
-      promptForUnsavedChanges,
-      setSelectedLayoutId,
-    ],
-  );
-
-  const onRenameLayout = useCallbackWithToast(
-    async (item: Layout, newName: string) => {
-      await layoutManager.updateLayout({ id: item.id, name: newName });
-      void analytics.logEvent(AppEvent.LAYOUT_RENAME, { permission: item.permission });
-    },
-    [analytics, layoutManager],
-  );
-
-  const onDuplicateLayout = useCallbackWithToast(
-    async (item: Layout) => {
-      if (state.selectedIds.length > 1) {
-        dispatch({ type: "queue-multi-action", action: "duplicate" });
-        return;
-      }
-
-      if (!(await promptForUnsavedChanges())) {
-        return;
-      }
-      const newLayout = await layoutManager.saveNewLayout({
-        name: `${item.name} copy`,
-        data: item.working?.data ?? item.baseline.data,
-        permission: "CREATOR_WRITE",
-      });
-      await onSelectLayout(newLayout);
-      void analytics.logEvent(AppEvent.LAYOUT_DUPLICATE, { permission: item.permission });
-    },
-    [
-      analytics,
-      dispatch,
-      layoutManager,
-      onSelectLayout,
-      promptForUnsavedChanges,
-      state.selectedIds.length,
-    ],
-  );
-
-  const onDeleteLayout = useCallbackWithToast(
-    async (item: Layout) => {
-      if (state.selectedIds.length > 1) {
-        dispatch({ type: "queue-multi-action", action: "delete" });
-        return;
-      }
-
-      void analytics.logEvent(AppEvent.LAYOUT_DELETE, { permission: item.permission });
-
-      // If the layout was selected, select a different available layout.
-      //
-      // When a users current layout is deleted, we display a notice. By selecting a new layout
-      // before deleting their current layout we avoid the weirdness of displaying a notice that the
-      // user just deleted their current layout which is somewhat obvious to the user.
-      if (currentLayoutId === item.id) {
-        const storedLayouts = await layoutManager.getLayouts();
-        const targetLayout = storedLayouts.find((layout) => layout.id !== currentLayoutId);
-        setSelectedLayoutId(targetLayout?.id);
-        dispatch({ type: "select-id", id: targetLayout?.id });
-      }
-      await layoutManager.deleteLayout({ id: item.id });
-    },
-    [
-      analytics,
-      currentLayoutId,
-      dispatch,
-      layoutManager,
-      setSelectedLayoutId,
-      state.selectedIds.length,
-    ],
-  );
-
   const createNewLayout = useCallbackWithToast(async () => {
     if (!(await promptForUnsavedChanges())) {
       return;
@@ -372,15 +219,7 @@ export default function LayoutBrowser({
     void analytics.logEvent(AppEvent.LAYOUT_CREATE);
   }, [promptForUnsavedChanges, currentDateForStorybook, layoutManager, onSelectLayout, analytics]);
 
-  const onExportLayout = useCallbackWithToast(
-    async (item: Layout) => {
-      const content = JSON.stringify(item.working?.data ?? item.baseline.data, undefined, 2) ?? "";
-      downloadTextFile(content, `${item.name}.json`);
-      void analytics.logEvent(AppEvent.LAYOUT_EXPORT, { permission: item.permission });
-    },
-    [analytics],
-  );
-
+  // function for a feature that we currently don't have?
   const onShareLayout = useCallbackWithToast(
     async (item: Layout) => {
       const name = await prompt({
@@ -402,46 +241,7 @@ export default function LayoutBrowser({
     [analytics, layoutManager, onSelectLayout, prompt],
   );
 
-  const onOverwriteLayout = useCallbackWithToast(
-    async (item: Layout) => {
-      // We don't need to confirm the multiple selection case because we force users to save
-      // or abandon changes before selecting another layout with unsaved changes to the current
-      // shared layout.
-      if (state.selectedIds.length > 1) {
-        dispatch({ type: "queue-multi-action", action: "save" });
-        return;
-      }
-
-      if (layoutIsShared(item)) {
-        const response = await confirm({
-          title: `Update “${item.name}”?`,
-          prompt:
-            "Your changes will overwrite this layout for all organization members. This cannot be undone.",
-          ok: "Save",
-        });
-        if (response !== "ok") {
-          return;
-        }
-      }
-      await layoutManager.overwriteLayout({ id: item.id });
-      void analytics.logEvent(AppEvent.LAYOUT_OVERWRITE, { permission: item.permission });
-    },
-    [analytics, confirm, dispatch, layoutManager, state.selectedIds.length],
-  );
-
-  const onRevertLayout = useCallbackWithToast(
-    async (item: Layout) => {
-      if (state.selectedIds.length > 1) {
-        dispatch({ type: "queue-multi-action", action: "revert" });
-        return;
-      }
-
-      await layoutManager.revertLayout({ id: item.id });
-      void analytics.logEvent(AppEvent.LAYOUT_REVERT, { permission: item.permission });
-    },
-    [analytics, dispatch, layoutManager, state.selectedIds.length],
-  );
-
+  // another funciton connected to layouts and organizations, something we dont use right now
   const onMakePersonalCopy = useCallbackWithToast(
     async (item: Layout) => {
       const newLayout = await layoutManager.makePersonalCopy({
@@ -456,78 +256,6 @@ export default function LayoutBrowser({
     },
     [analytics, layoutManager, onSelectLayout],
   );
-
-  const importLayout = useCallbackWithToast(async () => {
-    if (!(await promptForUnsavedChanges())) {
-      return;
-    }
-    const fileHandles = await showOpenFilePicker({
-      multiple: true,
-      excludeAcceptAllOption: false,
-      types: [
-        {
-          description: "JSON Files",
-          accept: {
-            "application/json": [".json"],
-          },
-        },
-      ],
-    });
-    if (fileHandles.length === 0) {
-      return;
-    }
-
-    const newLayouts = await Promise.all(
-      fileHandles.map(async (fileHandle) => {
-        const file = await fileHandle.getFile();
-        const layoutName = path.basename(file.name, path.extname(file.name));
-        const content = await file.text();
-
-        if (!isMounted()) {
-          return;
-        }
-
-        let parsedState: unknown;
-        try {
-          parsedState = JSON.parse(content);
-        } catch (err: unknown) {
-          enqueueSnackbar(`${file.name} is not a valid layout: ${(err as Error).message}`, {
-            variant: "error",
-          });
-          return;
-        }
-
-        if (typeof parsedState !== "object" || !parsedState) {
-          enqueueSnackbar(`${file.name} is not a valid layout`, { variant: "error" });
-          return;
-        }
-
-        const data = parsedState as LayoutData;
-        const newLayout = await layoutManager.saveNewLayout({
-          name: layoutName,
-          data,
-          permission: "CREATOR_WRITE",
-        });
-        return newLayout;
-      }),
-    );
-
-    if (!isMounted()) {
-      return;
-    }
-    const newLayout = newLayouts.find((layout) => layout != undefined);
-    if (newLayout) {
-      void onSelectLayout(newLayout);
-    }
-    void analytics.logEvent(AppEvent.LAYOUT_IMPORT, { numLayouts: fileHandles.length });
-  }, [
-    analytics,
-    enqueueSnackbar,
-    isMounted,
-    layoutManager,
-    onSelectLayout,
-    promptForUnsavedChanges,
-  ]);
 
   const [enableNewTopNav = true] = useAppConfigurationValue<boolean>(AppSetting.ENABLE_NEW_TOPNAV);
   const [hideSignInPrompt = false, setHideSignInPrompt] = useAppConfigurationValue<boolean>(
